server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.44.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.30.0",
    "langchain>=0.3.27",
    "langgraph>=0.6.5",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pyright>=1.1.403",
]
